refactor(api): reorganize STT providers into directory structure
Restructure apps/api/src/stt to follow the pattern from owhisper-client/src/adapter:
- Create provider directories: deepgram/, assemblyai/, soniox/
- Move live streaming code to <provider>/live.ts
- Move batch transcription code to <provider>/batch.ts
- Add index.ts for each provider to export both modules
- Update main index.ts to import from provider directories
