+++ apps/api/src/stt/assemblyai/index.ts
@@ -1,0 +1,2 @@
+export { buildAssemblyAIUrl, createAssemblyAIProxy } from "./live";
+export { transcribeWithAssemblyAI } from "./batch";
