+++ apps/api/src/stt/deepgram/index.ts
@@ -1,0 +1,2 @@
+export { buildDeepgramUrl, createDeepgramProxy } from "./live";
+export { transcribeWithDeepgram } from "./batch";
