+++ apps/api/src/stt/index.ts
@@ -1,21 +1,27 @@
-import { createAssemblyAIProxy } from "./assemblyai";
-import { transcribeWithAssemblyAI } from "./batch-assemblyai";
-import { transcribeWithDeepgram } from "./batch-deepgram";
-import { transcribeWithSoniox } from "./batch-soniox";
+import { createAssemblyAIProxy, transcribeWithAssemblyAI } from "./assemblyai";
 import type { BatchParams, BatchProvider, BatchResponse } from "./batch-types";
 import { WsProxyConnection } from "./connection";
-import { createDeepgramProxy } from "./deepgram";
-import { createSonioxProxy } from "./soniox";
+import { createDeepgramProxy, transcribeWithDeepgram } from "./deepgram";
+import { createSonioxProxy, transcribeWithSoniox } from "./soniox";
 
 export { WsProxyConnection, type WsProxyOptions } from "./connection";
 export { normalizeWsData, type WsPayload } from "./utils";
-export { buildDeepgramUrl, createDeepgramProxy } from "./deepgram";
-export { buildAssemblyAIUrl, createAssemblyAIProxy } from "./assemblyai";
-export { buildSonioxUrl, createSonioxProxy } from "./soniox";
+export {
+  buildDeepgramUrl,
+  createDeepgramProxy,
+  transcribeWithDeepgram,
+} from "./deepgram";
+export {
+  buildAssemblyAIUrl,
+  createAssemblyAIProxy,
+  transcribeWithAssemblyAI,
+} from "./assemblyai";
+export {
+  buildSonioxUrl,
+  createSonioxProxy,
+  transcribeWithSoniox,
+} from "./soniox";
 export type { BatchParams, BatchProvider, BatchResponse } from "./batch-types";
-export { transcribeWithDeepgram } from "./batch-deepgram";
-export { transcribeWithAssemblyAI } from "./batch-assemblyai";
-export { transcribeWithSoniox } from "./batch-soniox";
 
 export const UPSTREAM_URL_HEADER = "x-owh-upstream-url";
 export const UPSTREAM_AUTH_HEADER = "x-owh-upstream-auth";
